requirements.txt
redis
requests
pandas
numpy
python-dotenv
google-generativeai
crewai>=0.28.0
chromadb>=0.4.0
reportlab
streamlit